+++ src/module/setupModules.ts
@@ -2,7 +2,7 @@ import { debug, error, debugEnabled } from "../midi-qol.js";
 import { log } from "../midi-qol.js";
 
 let modules = {"about-time": "0.0", 
-              "betterrolls5e": "1.1.0", 
+              "betterrolls5e": "1.6.6", 
               "dice-so-nice": "4.1.1", 
               "itemacro": "1.0.0", 
               "lmrtfy": "0.9",