+++ Changelog.md
@@ -1,3 +1,13 @@
+## 0.8.43
+* Support better rolls rolls for saves/test/skills - requires better rolls 1.6.6. (Finally)
+* Fixed a bug introduced in 0.8.42 that swapped ability save/checks when auto rolling a save for a actor.
+* Fixed a bug for earlier versions of CUB that did not include an effect icon for concentration.
+* If Convenient Effects is enabld use the status effect from Convenient Effects preferentially to other choices.
+* Fix for spells with a range area effect to measure all squares covered by the potential target token.
+* Fix for special duration effects being removed as a result of the attack that created the effect (i.e. isAttacked etc).
+* Fix for missing targets with a lower AC than the hit roll. Some actors are being created with a string in ac.value, rather than a number, which was confusing midi-qol.
+* Modified the d20 attack roll + damage total to hide the damage formula from players.
+
 ## 0.8.42
 * Fix for template targeting with token magic fx.